+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `collections`
         - Components
             - `collection-subjects-list`
+            - `discover-page/facets/taxonomy`
 
 ## [19.7.1] - 2019-08-05
 ### Changed