+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
         - Components
             - `collection-subjects-list`
             - `discover-page/facets/taxonomy`
+            - `discover-page/active-filters/taxonomy`
 
 ## [19.7.1] - 2019-08-05
 ### Changed